floats and not doubles, compiler optimization settings

+++ Classes/Game/globals.h
@@ -85,6 +85,6 @@ namespace game
 
 extern game::GameState g_GameState;
 extern bool g_ParticlesEnabled;
-extern double g_FPS;
+extern float g_FPS;
